requirements.txt
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
Django==4.2.5
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
drf-extra-fields==3.7.0
filetype==1.2.0
future==0.18.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pycparser==2.21
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tornado==6.3.3
urllib3==2.0.5
